requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
exa-py>=0.6.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.4.2
pytest-asyncio>=0.21.1
